Added super simple permissions check for functionalities/views

+++ src/app/drones-list-view/drones-list-view.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Drone } from '../dto/drone.dto';
+import { Permissions } from '../dto/permissions.dto';
 
 import { DroneService } from '../Services/drone.service';
 import { FirebaseService } from '../Services/firebase.service';
@@ -14,6 +15,7 @@ export class DronesListViewComponent implements OnInit {
 
   drones : Drone[] = [];
   count:number;
+  permissions: Permissions = new Permissions(false, false, false, false);
 
   constructor(private droneService : DroneService, private firebaseService : FirebaseService, private router: Router) { }
 
@@ -23,6 +25,9 @@ export class DronesListViewComponent implements OnInit {
       this.router.navigate(['login']);
     }
 
+    let content = JSON.parse(localStorage.getItem('user'));
+    this.permissions = this.firebaseService.checkPermissionsForRole(content.email);
+
     this.droneService.getDrones().subscribe(drones => {
       this.drones = drones;
       this.count = this.drones.length;
@@ -31,9 +36,14 @@ export class DronesListViewComponent implements OnInit {
   }
 
   deleteDrone(event, drone){
-    if(confirm('Are you sure you want to delete the selected drone?')){
-      this.droneService.deleteDrone(drone);
-
+    if(this.permissions.hasDeleteAccess){
+      if(confirm('Are you sure you want to delete the selected drone?')){
+        this.droneService.deleteDrone(drone);
+      }
+    } else {
+        alert('Uh oh! It seems like you do not have permissions to delete drones..');
+        this.router.navigate(['home']);
     }
+   
   }
 }
